+++ src/transform/lower_l2_persistent_annotation.cc
@@ -10,7 +10,6 @@
 #include <tvm/tir/transform.h>
 
 #include "../op/builtin.h"
-#include "../op/bulk_copy.h"
 #include "../runtime/runtime.h"
 
 namespace tvm {
